requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
SQLAlchemy>=2.0
PyYAML>=6.0
pydantic>=2.6
